[FIX] l10n_br_account_export: planilha deterministica e helper unico
O xlsxwriter grava a hora de criacao dentro do arquivo, entao duas geracoes da
mesma exportacao produziam bytes diferentes e o determinismo exigido do produto
nao se sustentava em XLSX. A data de criacao passa a ser derivada do proprio
lote (a data final do periodo) em vez do relogio.

O teste de determinismo do contrato so denunciava a diferenca quando as duas
geracoes caiam em segundos distintos, o que o tornava intermitente: passou no
PR anterior e falhou neste. Agora ele espera de proposito entre as geracoes.

Aproveita para tirar a triplicacao: a montagem da planilha estava copiada nos
tres modulos de planilha e passa ao chassi, com import preguicoso do xlsxwriter
para o chassi seguir instalavel sem a biblioteca. Cada layout declara apenas as
suas colunas.
