+++ l10n_br_account_export/models/__init__.py
@@ -4,5 +4,6 @@
 from . import format_helper
 from . import account_export_config
 from . import account_export
+from . import sheet_helper
 from . import account_move
 from . import account_account
